requirements.txt
playwright
pytest
requests
asyncio
python-dotenv
pytest-mock